requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
plotly
requests
beautifulsoup4
psycopg2-binary
mysql-connector-python
